float2string #4 (Spacing)
git-svn-id: svn://svn.lyx.org/lyx/lyx-devel/trunk@9445 a592a061-630c-0410-9148-cb99ea01b6c8

+++ src/rowpainter.C
@@ -509,7 +509,7 @@ void RowPainter::paintFirst()
 			// the top.
 			if (layout->counter == "chapter") {
 				if (buffer.params().secnumdepth >= 0) {
-					float spacing_val = 1.0;
+					double spacing_val = 1.0;
 					if (!parparams.spacing().isDefault()) {
 						spacing_val = parparams.spacing().getValue();
 					} else {
@@ -550,7 +550,7 @@ void RowPainter::paintFirst()
 		LyXFont font = getLabelFont();
 		if (!par_.getLabelstring().empty()) {
 			string const str = par_.getLabelstring();
-			float spacing_val = 1.0;
+			double spacing_val = 1.0;
 			if (!parparams.spacing().isDefault()) {
 				spacing_val = parparams.spacing().getValue();
 			} else {
